rename Network to BitcoinNetwork create new Network interface
